# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tinder-Like API",
        "description": "API for Tinder-Like application",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/people/recommendations": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "Get people recommendations",
                "operationId": "8523267be2fe11fe4d5b2411dffcce74",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/people/{personId}/liked": {
            "get": {
                "tags": [
                    "Likes"
                ],
                "summary": "Get people liked by a user",
                "operationId": "b02c06c3fff7b4be10fdca58b793b557",
                "parameters": [
                    {
                        "name": "personId",
                        "in": "path",
                        "description": "ID of the person",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/people/{personId}/like/{likedPersonId}": {
            "post": {
                "tags": [
                    "Likes"
                ],
                "summary": "Like a person",
                "operationId": "0b6ed94f56d2466592dd64e2b0791363",
                "parameters": [
                    {
                        "name": "personId",
                        "in": "path",
                        "description": "ID of the person who is liking",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "likedPersonId",
                        "in": "path",
                        "description": "ID of the person being liked",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Person liked successfully"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cannot like yourself"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/people/{personId}/dislike/{dislikedPersonId}": {
            "post": {
                "tags": [
                    "Likes"
                ],
                "summary": "Dislike a person",
                "operationId": "3444c5deb84c805300b8b486803663ec",
                "parameters": [
                    {
                        "name": "personId",
                        "in": "path",
                        "description": "ID of the person who is disliking",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "dislikedPersonId",
                        "in": "path",
                        "description": "ID of the person being disliked",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Person disliked successfully"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cannot dislike yourself"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Error": {
                "required": [
                    "message",
                    "success"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "success": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "People",
            "description": "People"
        },
        {
            "name": "Likes",
            "description": "Likes"
        }
    ]
}